Optimize file descriptor allocation using a hint
Replaces the O(N) linear scan for free file descriptors with a O(1)
hint-based approach. A `next_fd` hint is added to the process structure,
which tracks the lowest possible free FD.

Also includes necessary build fixes to support the current toolchain:
- Updated Makefile to use -std=c2x for GCC 13 compatibility
- Fixed pointer sign mismatches in VFS read/write operations
- Added 64-bit division helpers for freestanding environment
- Resolved linker errors in PMM and personality subsystems
- Corrected include paths for storage drivers

+++ sys/fs/fat/fat.c
@@ -177,7 +177,7 @@ static struct dirent *fat_readdir(fs_node_t *node, uint64_t index) {
             fat_dirent_t *entry = (fat_dirent_t *)(dir_buf + i);
             
             if (entry->name[0] == 0x00) return NULL;  // End of directory
-            if (entry->name[0] == 0xE5) continue;      // Deleted entry
+            if ((uint8_t)entry->name[0] == 0xE5) continue;      // Deleted entry
             
             // Check for LFN entry
             if (entry->attr == FAT_ATTR_LFN) {
@@ -267,7 +267,7 @@ static fs_node_t *fat_finddir(fs_node_t *node, char *name) {
             fat_dirent_t *entry = (fat_dirent_t *)(dir_buf + i);
             
             if (entry->name[0] == 0x00) return NULL;  // End of directory
-            if (entry->name[0] == 0xE5) continue;      // Deleted entry
+            if ((uint8_t)entry->name[0] == 0xE5) continue;      // Deleted entry
             
             // Check for LFN entry
             if (entry->attr == FAT_ATTR_LFN) {
